refactor(ui): move deriveSyncViewModel into its own module, barrel the view-model
Finish the view-model.ts decomposition. deriveSyncViewModel (and its
private helpers MergedDevice, createRepairItem, createReviewItem,
createNamingItem, mergeDevices, isOfflineTeamDevice) move into
view-model/sync-view-model.ts. view-model.ts becomes a barrel that
re-exports the nine public functions plus all types so call sites in
the sync tabs, state.ts, and view-model.test.ts stay on the same
import path.
